mem_watcher:删除冗余代码
