--- eBPF_Supermarket/Memory_Subsystem/applications/mem_watcher/paf.h
@@ -1,59 +1,0 @@
-/* SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause) */
-/* Copyright (c) 2020 Facebook */
-#ifndef __PAF_H
-#define __PAF_H
-
-#define TASK_COMM_LEN 16
-#define MAX_FILENAME_LEN 127
-
-#define ___GFP_DMA              0x01u
-#define ___GFP_HIGHMEM          0x02u
-#define ___GFP_DMA32            0x04u
-#define ___GFP_MOVABLE          0x08u
-#define ___GFP_RECLAIMABLE      0x10u
-#define ___GFP_HIGH             0x20u
-#define ___GFP_IO               0x40u
-#define ___GFP_FS               0x80u
-#define ___GFP_WRITE            0x100u
-#define ___GFP_NOWARN           0x200u
-#define ___GFP_RETRY_MAYFAIL    0x400u
-#define ___GFP_NOFAIL           0x800u
-#define ___GFP_NORETRY          0x1000u
-#define ___GFP_MEMALLOC         0x2000u
-#define ___GFP_COMP             0x4000u
-#define ___GFP_ZERO             0x8000u
-#define ___GFP_NOMEMALLOC       0x10000u
-#define ___GFP_HARDWALL         0x20000u
-#define ___GFP_THISNODE         0x40000u
-#define ___GFP_ATOMIC           0x80000u
-#define ___GFP_ACCOUNT          0x100000u
-#define ___GFP_DIRECT_RECLAIM   0x200000u
-#define ___GFP_KSWAPD_RECLAIM   0x400000u
-	
-#define GFP_ATOMIC      (__GFP_HIGH|__GFP_ATOMIC|__GFP_KSWAPD_RECLAIM)
-#define GFP_KERNEL      (__GFP_RECLAIM | __GFP_IO | __GFP_FS)
-#define GFP_KERNEL_ACCOUNT (GFP_KERNEL | __GFP_ACCOUNT)
-#define GFP_NOWAIT      (__GFP_KSWAPD_RECLAIM)
-#define GFP_NOIO        (__GFP_RECLAIM)
-#define GFP_NOFS        (__GFP_RECLAIM | __GFP_IO)
-#define GFP_USER        (__GFP_RECLAIM | __GFP_IO | __GFP_FS | __GFP_HARDWALL)
-#define GFP_DMA         __GFP_DMA
-#define GFP_DMA32       __GFP_DMA32
-#define GFP_HIGHUSER    (GFP_USER | __GFP_HIGHMEM)
-#define GFP_HIGHUSER_MOVABLE    (GFP_HIGHUSER | __GFP_MOVABLE)
-#define GFP_TRANSHUGE_LIGHT     ((GFP_HIGHUSER_MOVABLE | __GFP_COMP | \
-                         __GFP_NOMEMALLOC | __GFP_NOWARN) & ~__GFP_RECLAIM)
-#define GFP_TRANSHUGE   (GFP_TRANSHUGE_LIGHT | __GFP_DIRECT_RECLAIM)
-
-struct paf_event {
-	unsigned long min;
-	unsigned long low;
-	unsigned long high;
-	unsigned long present;
-	unsigned long protection;
-	int flag;
-};
-
-
-
-#endif /* __PAF_H */
