+++ eBPF_Supermarket/Memory_Subsystem/applications/mem_watcher/pr.bpf.c
@@ -11,7 +11,7 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_tracing.h>
 #include <bpf/bpf_core_read.h>
-#include "pr.h"
+#include "mem_watcher.h"
 
 char LICENSE[] SEC("license") = "Dual BSD/GPL";
