+++ eBPF_Supermarket/Memory_Subsystem/applications/mem_watcher/sysstat.bpf.c
@@ -11,7 +11,7 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_tracing.h>
 #include <bpf/bpf_core_read.h>
-#include "sysstat.h"
+#include "mem_watcher.h"
 
 char LICENSE[] SEC("license") = "Dual BSD/GPL";
